Variant objects can now be used inside the TensorFlow execution graph.
Includes modifications and bugfixes to Variant and Tensor objects.

The most significant change is that a Variant storing a
VariantTensorDataProto acts like the data type inside the stored proto.

This means that its TypeName() is the typename of the pre-serialized data
object, and calling .get<OriginalType> performs a Decode under the hood.
Calling .get<OriginalType> on a **const** Variant of VariantTensorDataProto
causes a CHECK failure.

Also added proper TypeName() for unboxed primitive types (int, ...).
Change: 165444413